fix: remove remaining position as i32 cast in workspace.rs
Change Note.position from i32 to f64 to match Operation.position (f64)
and the DB schema (REAL). Remove all bridging casts: `position as i32`
in note_from_row_tuple and five `note.position as f64` casts when
building Operation/RetractInverse structs. Update integer literals in
AddPosition arm and Note constructors throughout workspace.rs, note.rs,
scripting/mod.rs, and scripting/display_helpers.rs. Fix test assertions
to use f64 literals (0.0, 1.0, etc.) and Vec<f64> with sort_by.

All 362 tests pass.

+++ krillnotes-core/src/core/note.rs
@@ -47,8 +47,8 @@ pub struct Note {
     pub node_type: String,
     /// ID of the parent note, or `None` for root-level notes.
     pub parent_id: Option<String>,
-    /// Zero-based sort order among siblings that share the same `parent_id`.
-    pub position: i32,
+    /// Fractional sort order among siblings that share the same `parent_id`.
+    pub position: f64,
     /// Unix timestamp (seconds) when this note was created.
     pub created_at: i64,
     /// Unix timestamp (seconds) of the most recent modification.
@@ -78,7 +78,7 @@ mod tests {
             title: "Test Note".to_string(),
             node_type: "TextNote".to_string(),
             parent_id: None,
-            position: 0,
+            position: 0.0,
             created_at: 1234567890,
             modified_at: 1234567890,
             created_by: 0,
